# Use a custom comparator over sort-by