--- a/index.js
+++ b/index.js
@@ -4,9 +4,10 @@ var xtend = require('xtend');
 var ShelfPack = require('shelf-pack');
 var queue = require('queue-async');
 var emptyPNG = new mapnik.Image(1, 1).encodeSync('png');
-var sortBy = require('sort-by');
 
-var heightAscThanNameComparator = sortBy('-height', 'id');
+function heightAscThanNameComparator(a, b) {
+    return (b.height - a.height) || (a.id < b.id ? -1 : 1);
+};
 
 /**
  * Pack a list of images with width and height into a sprite layout.