A little hand-timing shows that in `master` the time breakdown of `generateLayout` looks like this:

```
png: 120ms
sort: 196ms
pack: 4ms
```

The time for `svg => png` rasterization + encoding makes sense but :flushed: on the sort...

This branch

- Adds a basic benchmark for `generateLayout`
- Switches from `sort-by` to a custom comparator meant to do the same order comparison

**Before**

```
# generateLayout bench (concurrency=1,x10)
ok 5 null
ok 6 3001ms
# generateLayout bench (concurrency=4,x20)
ok 7 null
ok 8 5079ms
```

**After**

```
# generateLayout bench (concurrency=1,x10)
ok 5 null
ok 6 1478ms
# generateLayout bench (concurrency=4,x20)
ok 7 null
ok 8 2337ms
```

cc @jakepruitt @bhousel @ianshward